clays-be/requirements.txt
Django>=3.2.12
djangorestframework>=3.11.2
django-cors-headers==3.10.1

